updated get stories
