+++ stories/components/agenda.stories.js
@@ -9,9 +9,9 @@ import { html } from 'lit-element';
 import { withA11y } from '@storybook/addon-a11y';
 import { withKnobs } from '@storybook/addon-knobs';
 import { withWebComponentsKnobs } from 'storybook-addon-web-components-knobs';
-import { withSignIn } from '../.storybook/addons/signInAddon/signInAddon';
-import { withCodeEditor } from '../.storybook/addons/codeEditorAddon/codeAddon';
-import '../dist/es6/components/mgt-agenda/mgt-agenda';
+import { withSignIn } from '../../.storybook/addons/signInAddon/signInAddon';
+import { withCodeEditor } from '../../.storybook/addons/codeEditorAddon/codeAddon';
+import '../../dist/es6/components/mgt-agenda/mgt-agenda';
 
 export default {
   title: 'Components | mgt-agenda',
