+++ stories/components/login.stories.js
@@ -9,17 +9,22 @@ import { html } from 'lit-element';
 import { withA11y } from '@storybook/addon-a11y';
 import { withKnobs } from '@storybook/addon-knobs';
 import { withWebComponentsKnobs } from 'storybook-addon-web-components-knobs';
-import { withSignIn } from '../.storybook/addons/signInAddon/signInAddon';
-import { withCodeEditor } from '../.storybook/addons/codeEditorAddon/codeAddon';
-import '../dist/es6/components/mgt-people/mgt-people';
+import { withSignIn } from '../../.storybook/addons/signInAddon/signInAddon';
+import { withCodeEditor } from '../../.storybook/addons/codeEditorAddon/codeAddon';
+import '../../dist/es6/components/mgt-login/mgt-login';
 
 export default {
-  title: 'Components | mgt-people',
-  component: 'mgt-people',
+  title: 'Components | mgt-login',
+  component: 'mgt-login',
   decorators: [withA11y, withSignIn, withCodeEditor],
-  parameters: { options: { selectedPanel: 'storybookjs/knobs/panel' } }
+  parameters: {
+    options: { selectedPanel: 'mgt/sign-in' },
+    signInAddon: {
+      test: 'test'
+    }
+  }
 };
 
-export const People = () => html`
-  <mgt-people show-max="5"></mgt-people>
+export const Login = () => html`
+  <mgt-login></mgt-login>
 `;
