+++ stories/components/people.stories.js
@@ -9,17 +9,17 @@ import { html } from 'lit-element';
 import { withA11y } from '@storybook/addon-a11y';
 import { withKnobs } from '@storybook/addon-knobs';
 import { withWebComponentsKnobs } from 'storybook-addon-web-components-knobs';
-import { withSignIn } from '../.storybook/addons/signInAddon/signInAddon';
-import { withCodeEditor } from '../.storybook/addons/codeEditorAddon/codeAddon';
-import '../dist/es6/components/mgt-person/mgt-person';
+import { withSignIn } from '../../.storybook/addons/signInAddon/signInAddon';
+import { withCodeEditor } from '../../.storybook/addons/codeEditorAddon/codeAddon';
+import '../../dist/es6/components/mgt-people/mgt-people';
 
 export default {
-  title: 'Components | mgt-person',
-  component: 'mgt-person',
+  title: 'Components | mgt-people',
+  component: 'mgt-people',
   decorators: [withA11y, withSignIn, withCodeEditor],
   parameters: { options: { selectedPanel: 'storybookjs/knobs/panel' } }
 };
 
-export const person = () => html`
-  <mgt-person person-query="me" show-name show-email></mgt-person>
+export const People = () => html`
+  <mgt-people show-max="5"></mgt-people>
 `;
