+++ stories/components/person.stories.js
@@ -9,22 +9,17 @@ import { html } from 'lit-element';
 import { withA11y } from '@storybook/addon-a11y';
 import { withKnobs } from '@storybook/addon-knobs';
 import { withWebComponentsKnobs } from 'storybook-addon-web-components-knobs';
-import { withSignIn } from '../.storybook/addons/signInAddon/signInAddon';
-import { withCodeEditor } from '../.storybook/addons/codeEditorAddon/codeAddon';
-import '../dist/es6/components/mgt-login/mgt-login';
+import { withSignIn } from '../../.storybook/addons/signInAddon/signInAddon';
+import { withCodeEditor } from '../../.storybook/addons/codeEditorAddon/codeAddon';
+import '../../dist/es6/components/mgt-person/mgt-person';
 
 export default {
-  title: 'Components | mgt-login',
-  component: 'mgt-login',
+  title: 'Components | mgt-person',
+  component: 'mgt-person',
   decorators: [withA11y, withSignIn, withCodeEditor],
-  parameters: {
-    options: { selectedPanel: 'mgt/sign-in' },
-    signInAddon: {
-      test: 'test'
-    }
-  }
+  parameters: { options: { selectedPanel: 'storybookjs/knobs/panel' } }
 };
 
-export const Login = () => html`
-  <mgt-login></mgt-login>
+export const person = () => html`
+  <mgt-person person-query="me" show-name show-email></mgt-person>
 `;
