+++ stories/samples/general.stories.js
@@ -6,11 +6,11 @@
  */
 
 import { html } from 'lit-element';
-import { withSignIn } from '../.storybook/addons/signInAddon/signInAddon';
-import { withCodeEditor } from '../.storybook/addons/codeEditorAddon/codeAddon';
+import { withSignIn } from '../../.storybook/addons/signInAddon/signInAddon';
+import { withCodeEditor } from '../../.storybook/addons/codeEditorAddon/codeAddon';
 
 export default {
-  title: 'Samples | Simple',
+  title: 'Samples | General',
   component: 'mgt-combo',
   decorators: [withSignIn, withCodeEditor],
   parameters: {
